+ created buttons.js isolated from elements.js
