+++ FINANCE-TRACKER/src/elements.js
@@ -1,4 +1,3 @@
-import modal, { showModal } from "./modal";
 
 export function element() {
     const element = document.createElement('div')
@@ -24,21 +23,3 @@ export function expenseElement(info, value) {
     return expense;
 }
 
-export function buttonElement() {
-    const button = document.createElement('button')
-    button.setAttribute('class', 'button')
-
-    return button
-}
-
-export function expenseButton() {
-    const button = buttonElement();
-    button.setAttribute('class', 'expenseButton functionButton')
-
-    button.addEventListener('click', () => {
-        const modalWindow = modal();
-        showModal(modalWindow);
-    })
-
-    return button;
-}
