+++ FINANCE-TRACKER/src/modal.js
@@ -1,8 +1,13 @@
 export default function modal() {
+    const modalBG = document.createElement('div');
+    modalBG.setAttribute('class', 'modalBG');
+
     const modal = document.createElement('div');
     modal.setAttribute('class', 'modal');
 
-    return modal
+    modalBG.appendChild(modal);
+
+    return modalBG;
 }
 
 export function showModal(modal) {
