+++ FINANCE-TRACKER/src/styles/modal.css
@@ -9,3 +9,13 @@
 
     background-color: #1a1a1a;
 }
+
+.modalBG {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+
+    background-color: rgba(0, 0, 0, 0.5);
+}
